+++ src/components/Header/components/StatsDialog.styles.ts
@@ -1,0 +1,11 @@
+import { styled } from '@mui/material/styles';
+import Dialog from '@mui/material/Dialog';
+
+export const BootstrapDialog = styled(Dialog)(({ theme }) => ({
+  '& .MuiDialogContent-root': {
+    padding: theme.spacing(2),
+  },
+  '& .MuiDialogActions-root': {
+    padding: theme.spacing(1),
+  },
+}));
