+++ src/components/Header/components/StatsDialog.tsx
@@ -1,7 +1,5 @@
 import { useState } from 'react';
 import Button from '@mui/material/Button';
-import { styled } from '@mui/material/styles';
-import Dialog from '@mui/material/Dialog';
 import DialogTitle from '@mui/material/DialogTitle';
 import DialogContent from '@mui/material/DialogContent';
 import DialogActions from '@mui/material/DialogActions';
@@ -9,18 +7,10 @@ import IconButton from '@mui/material/IconButton';
 import CloseIcon from '@mui/icons-material/Close';
 import Typography from '@mui/material/Typography';
 import { IoAnalyticsOutline } from 'react-icons/io5';
-import { useAllBlocks } from '../../api/hooks/useBlocks';
-import { toHexString, shortenString } from '../../utils/helpers';
-import CopyToClipboard from '../../components/CopyToClipboard';
-
-const BootstrapDialog = styled(Dialog)(({ theme }) => ({
-  '& .MuiDialogContent-root': {
-    padding: theme.spacing(2),
-  },
-  '& .MuiDialogActions-root': {
-    padding: theme.spacing(1),
-  },
-}));
+import { useAllBlocks } from '../../../api/hooks/useBlocks';
+import { toHexString, shortenString } from '../../../utils/helpers';
+import CopyToClipboard from '../../../components/CopyToClipboard';
+import { BootstrapDialog } from './StatsDialog.styles';
 
 function StatsDialog() {
   const [open, setOpen] = useState(false);
