+++ src/components/Header/components/StatsItem.tsx
@@ -1,0 +1,20 @@
+import {
+  StyledBox,
+  ValueTypography,
+  LabelTypography,
+} from './StatsItem.styles';
+
+interface Props {
+  label: string;
+  value: string;
+  lowerCase?: boolean;
+}
+
+export default function StatsItem({ label, value, lowerCase }: Props) {
+  return (
+    <StyledBox>
+      <ValueTypography lowerCase={lowerCase}>{value}</ValueTypography>
+      <LabelTypography>{label}</LabelTypography>
+    </StyledBox>
+  );
+}
